+++ src/Datadog.Trace.ClrProfiler.Managed/Emit/DynamicMethodBuilder.cs
@@ -6,13 +6,13 @@ using System.Reflection;
 using System.Reflection.Emit;
 using Sigil;
 
-namespace Datadog.Trace.ClrProfiler
+namespace Datadog.Trace.ClrProfiler.Emit
 {
     /// <summary>
     /// Helper class to instances of <see cref="DynamicMethod"/> using <see cref="System.Reflection.Emit"/>.
     /// </summary>
     /// <typeparam name="TDelegate">The type of delegate</typeparam>
-    public static class DynamicMethodBuilder<TDelegate>
+    internal static class DynamicMethodBuilder<TDelegate>
         where TDelegate : Delegate
     {
         private static readonly ConcurrentDictionary<Key, TDelegate> _cached = new ConcurrentDictionary<Key, TDelegate>(new KeyComparer());
